+++ src/pdi_migration/generator/__init__.py
@@ -1,3 +1,4 @@
+from pdi_migration.generator.kjb import KjbGenerator
 from pdi_migration.generator.ktr import KtrGenerator
 
-__all__ = ["KtrGenerator"]
+__all__ = ["KjbGenerator", "KtrGenerator"]